+++ python/cudf/io/csv.py
@@ -38,7 +38,7 @@ def read_csv(filepath_or_buffer, lineterminator='\n',
              thousands=None, decimal='.', true_values=None, false_values=None,
              nrows=None, byte_range=None, skip_blank_lines=True, comment=None,
              na_values=None, keep_default_na=True, na_filter=True,
-             prefix=None):
+             prefix=None, index_col=None):
 
     """
     Load and parse a CSV file into a DataFrame
@@ -121,6 +121,8 @@ def read_csv(filepath_or_buffer, lineterminator='\n',
         Passing False can improve performance.
     prefix : str, default None
         Prefix to add to column numbers when parsing without a header row
+    index_col : int or string, default None
+        Column to use as the row labels
 
     Returns
     -------
@@ -367,6 +369,13 @@ def read_csv(filepath_or_buffer, lineterminator='\n',
     for k, v in zip(new_names, outcols):
         df[k] = v
 
+    # Set index if the index_col parameter is passed
+    if index_col is not None and index_col is not False:
+        if isinstance(index_col, (int)):
+            df = df.set_index(df.columns[index_col])
+        else:
+            df = df.set_index(index_col)
+
     nvtx_range_pop()
 
     return df
@@ -382,7 +391,7 @@ def read_csv_strings(filepath_or_buffer, lineterminator='\n',
                      true_values=None, false_values=None, nrows=None,
                      byte_range=None, skip_blank_lines=True, comment=None,
                      na_values=None, keep_default_na=True, na_filter=True,
-                     prefix=None):
+                     prefix=None, index_col=None):
 
     """
     **Experimental**: This function exists only as a beta way to use