+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - PR #988 Add top-level merge function to cuDF
 - PR #992 Add comparison binaryops to DateTime columns
 - PR #996 Replace relative path imports with absolute paths in tests
+- PR #995 CSV Reader: Add index_col parameter to specify the column name or index to be used as row labels
 
 ## Improvements